:recycle: Convert re-order-flight-segments-api to TypeScript.
